add/ verif pour offer terminé

+++ CesiTonStage/app/Http/Controllers/OfferController.php
@@ -23,6 +23,12 @@ class OfferController extends Controller
 
     public function create()
     {
+
+        if (!session('account') || (int) session('account')->Id_Role == 1) {
+            // Rediriger l'utilisateur avec un message d'erreur (il n'est pas censé s'afficher car il y a en amond un bloquage visuel)
+            return back()->with('error', "Vous ne pouvez pas créer d'offre, vous n'en avez pas la permission");
+        }
+
         $accounts = Account::whereIn('Id_Role', [2, 3])->get();
         $companies = Company::all();
         $statuses = Status::whereIn('Id_Status', [1])->get();
@@ -61,6 +67,12 @@ class OfferController extends Controller
 
     public function edit($id)
     {
+
+        if (!session('account') || (int) session('account')->Id_Role == 1) {
+            // Rediriger l'utilisateur avec un message d'erreur (il n'est pas censé s'afficher car il y a en amond un bloquage visuel)
+            return back()->with('error', "Vous ne pouvez modifier cette offre, vous n'avez pas la permission pour.");
+        }
+
         $offer = Offer::findOrFail($id);
         return view('offers.edit', compact('offer'));
     }
@@ -86,6 +98,13 @@ class OfferController extends Controller
 
     public function destroy($id)
     {
+
+        if (!session('account') || (int) session('account')->Id_Role == 1) {
+            // Rediriger l'utilisateur avec un message d'erreur (il n'est pas censé s'afficher car il y a en amond un bloquage visuel)
+            return back()->with('error', "Vous ne pouvez pas supprimer cette offre, vous n'avez pas la permission");
+        }
+
+
         $offer = Offer::findOrFail($id);
         $offer->delete();
         return redirect()->route('offers.index')->with('success', 'Offre supprimée avec succès.');
